feat: update repositories to implement search and findAll methods for people; refactor better_auth_id to be unique in models
