+++ prisma/schema.prisma
@@ -268,7 +268,7 @@ model school {
 model parent {
   id               String          @id @default(dbgenerated("gen_random_uuid()")) @map("_id") @db.Uuid
   branch_id        String          @db.Uuid
-  better_auth_id   String          @db.VarChar(255)
+  better_auth_id   String          @unique @db.VarChar(255)
   first_name       String          @db.VarChar(100)
   middle_name      String          @db.VarChar(100)
   phone            String          @db.VarChar(50)
@@ -278,12 +278,13 @@ model parent {
   note             String?
   studentRelations parentstudent[]
   branch           branch?         @relation(fields: [branch_id], references: [id], onDelete: Cascade, onUpdate: Cascade)
+  user             User?           @relation(fields: [better_auth_id], references: [id], onDelete: Cascade, onUpdate: Cascade)
 }
 
 model staff {
   id             String          @id @default(dbgenerated("gen_random_uuid()")) @map("_id") @db.Uuid
   branch_id      String?         @db.Uuid
-  better_auth_id String          @db.VarChar(255)
+  better_auth_id String          @unique @db.VarChar(255)
   first_name     String          @db.VarChar(100)
   middle_name    String          @db.VarChar(100)
   phone          String?         @db.VarChar(50)
@@ -291,12 +292,13 @@ model staff {
   sex            sex_enum?
   role           staff_role_enum
   branch         branch?         @relation(fields: [branch_id], references: [id], onDelete: Cascade, onUpdate: Cascade)
+  user           User?           @relation(fields: [better_auth_id], references: [id], onDelete: Cascade, onUpdate: Cascade)
 }
 
 model student {
   id                          String       @id @default(dbgenerated("gen_random_uuid()")) @map("_id") @db.Uuid
   branch_id                   String       @db.Uuid
-  better_auth_id              String       @db.VarChar(255)
+  better_auth_id              String       @unique @db.VarChar(255)
   first_name                  String       @db.VarChar(100)
   middle_name                 String       @db.VarChar(100)
   last_name                   String       @db.VarChar(100)
@@ -311,11 +313,12 @@ model student {
   student_registration_number Int          @default(autoincrement())
   enrollment                  enrollment[]
   branch                      branch?      @relation(fields: [branch_id], references: [id], onDelete: Cascade, onUpdate: Cascade)
+  user                        User?        @relation(fields: [better_auth_id], references: [id], onDelete: Cascade, onUpdate: Cascade)
 }
 
 model teacher {
   id                          String            @id @default(dbgenerated("gen_random_uuid()")) @map("_id") @db.Uuid
-  better_auth_id              String            @db.VarChar(255)
+  better_auth_id              String            @unique @db.VarChar(255)
   first_name                  String            @db.VarChar(100)
   middle_name                 String            @db.VarChar(100)
   phone                       String            @db.VarChar(50)
@@ -328,6 +331,7 @@ model teacher {
   branch_id                   String            @db.Uuid
   libraryItemLoan             libraryitemloan[]
   branch                      branch?           @relation(fields: [branch_id], references: [id], onDelete: Cascade, onUpdate: Cascade)
+  user                        User?             @relation(fields: [better_auth_id], references: [id], onDelete: Cascade, onUpdate: Cascade)
 
   timetable         timetable[]
   assessment        assessment[]
@@ -513,6 +517,10 @@ model User {
   banExpires    DateTime?
   sessions      Session[]
   accounts      Account[]
+  teacher       teacher?
+  students      student?
+  staffs        staff?
+  parents       parent?
 
   @@unique([email])
   @@map("user")
