+++ src/repositories/staff.repositorie.ts
@@ -1,4 +1,4 @@
-import { Response } from "express";
+import { Response, Request } from "express";
 import { BaseRepository } from "./base.repositorie.js";
 
 export class staffRepo extends BaseRepository<"staff"> {
@@ -12,5 +12,18 @@ export class staffRepo extends BaseRepository<"staff"> {
 
   async delete(id: string, signal?: AbortSignal) {
     return this.delete_people(id, signal);
+  }
+
+  async search(req: Request) {
+    return this.search_people(req);
+  }
+
+  async findAll(options: {
+    where?: any;
+    orderBy?: any;
+    take?: number;
+    signal?: AbortSignal;
+  }) {
+    return this.findAll_people(options);
   }
 }
